+++ src/AnalogLimit.cpp
@@ -5,7 +5,7 @@ AnalogLimit::AnalogLimit() {
   // Initialisiert alle Sensoren mit einem konservativen Standardwert (888).
   // Dies verhindert Fehltriggerungen, falls vor der Kalibrierung ausgelesen
   // wird.
-  for (int i = 0; i < NUM_SPOTS; i++) {
+  for (uint8_t i = 0; i < NUM_SPOTS; i++) {
     limits[i] = 888;
   }
 }
@@ -14,11 +14,11 @@ void AnalogLimit::calibrate() {
   // Führt eine Nullpunktkalibrierung für alle Stellplätze durch.
   // Es wird davon ausgegangen, dass sich beim Start KEINE Becher auf den
   // Stellplätzen befinden.
-  for (int i = 0; i < NUM_SPOTS; i++) {
-    int temp = 0;
+  for (uint8_t i = 0; i < NUM_SPOTS; i++) {
+    uint16_t temp = 0;
 
     // 10 Messungen durchführen, um Rauschen zu minimieren
-    for (int j = 0; j < 10; j++) {
+    for (uint8_t j = 0; j < 10; j++) {
       temp += analogRead(PIN_SENSORS[i]);
       delay(10);
     }
@@ -31,4 +31,4 @@ void AnalogLimit::calibrate() {
   }
 }
 
-int AnalogLimit::getValue(int position) const { return limits[position]; }
+uint16_t AnalogLimit::getValue(uint8_t position) const { return limits[position]; }
